✨ feat(middleware.ts): add authMiddleware and config to protect routes and allow customization
The `middleware.ts` file is added to the project. It includes the `authMiddleware` function from the `@clerk/nextjs` package. This middleware is used to protect all routes, including `api/trpc` routes. The `config` object is also defined to specify the routes that should be protected. This allows for customization of which routes should be public and which should be protected. For more information about configuring the middleware, please refer to the [Clerk documentation](https://clerk.com/docs/nextjs/middleware).